pyproject.toml
[project]
name = "empathyengine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "elevenlabs>=2.43.0",
    "google-cloud-texttospeech>=2.36.0",
    "gtts>=2.5.4",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "streamlit>=1.56.0",
    "torch>=2.11.0",
    "torchaudio>=2.11.0",
    "transformers>=5.5.4",
]
